Reviewed-on: https://gitea.obmondo.com/EnableIT/LinuxAid/pulls/1191

+++ modules/upstream/inifile/lib/puppet/type/ini_setting.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'digest/md5'
 require 'puppet/parameter/boolean'
 
@@ -45,9 +47,7 @@ Puppet::Type.newtype(:ini_setting) do
   newparam(:setting) do
     desc 'The name of the setting to be defined.'
     munge do |value|
-      if value.match?(%r{(^\s|\s$)})
-        Puppet.warn('Settings should not have spaces in the value, we are going to strip the whitespace')
-      end
+      Puppet.warn('Settings should not have spaces in the value, we are going to strip the whitespace') if value.match?(%r{(^\s|\s$)})
       value.strip
     end
   end
@@ -60,9 +60,7 @@ Puppet::Type.newtype(:ini_setting) do
   newparam(:path) do
     desc 'The ini file Puppet will ensure contains the specified setting.'
     validate do |value|
-      unless Puppet::Util.absolute_path?(value)
-        raise(Puppet::Error, _("File paths must be fully qualified, not '%{value}'") % { value: value })
-      end
+      raise(Puppet::Error, _("File paths must be fully qualified, not '%{value}'") % { value: value }) unless Puppet::Util.absolute_path?(value)
     end
   end
 
@@ -87,6 +85,7 @@ Puppet::Type.newtype(:ini_setting) do
     desc 'The value of the setting to be defined.'
 
     munge do |value|
+      value = value.unwrap if value.respond_to?(:unwrap)
       if ([true, false].include? value) || value.is_a?(Numeric)
         value.to_s
       else
@@ -98,13 +97,13 @@ Puppet::Type.newtype(:ini_setting) do
       if @resource[:show_diff] == :true && Puppet[:show_diff]
         newvalue
       elsif @resource[:show_diff] == :md5 && Puppet[:show_diff]
-        '{md5}' + Digest::MD5.hexdigest(newvalue.to_s)
+        "{md5}#{Digest::MD5.hexdigest(newvalue.to_s)}"
       else
         '[redacted sensitive information]'
       end
     end
 
-    def is_to_s(value) # rubocop:disable Style/PredicateName : Changing breaks the code (./.bundle/gems/gems/puppet-5.3.3-universal-darwin/lib/puppet/parameter.rb:525:in `to_s')
+    def is_to_s(value) # rubocop:disable Naming/PredicateName : Changing breaks the code (./.bundle/gems/gems/puppet-5.3.3-universal-darwin/lib/puppet/parameter.rb:525:in `to_s')
       should_to_s(value)
     end
 
@@ -142,9 +141,8 @@ Puppet::Type.newtype(:ini_setting) do
   end
 
   def refresh
-    if self[:ensure] == :absent && self[:refreshonly]
-      return provider.destroy
-    end
+    return provider.destroy if self[:ensure] == :absent && self[:refreshonly]
+
     # update the value in the provider, which will save the value to the ini file
     provider.value = self[:value] if self[:refreshonly]
   end
